requirements.txt
streamlit
openai
spacy
PyPDF2
python-docx
pandas
plotly
python-dotenv
tiktoken
reportlab